pyproject.toml
[project]
name = "openclaw-ai-quant"
version = "0.0.0"
description = "OpenClaw AI Quant trading engine (local)"
requires-python = ">=3.12"
dependencies = [
  "eth-account==0.13.7",
  "hyperliquid-python-sdk==0.21.0",
  "pandas==3.0.0",
  "PyYAML==6.0.3",
  "ta==0.11.0",
  "websocket-client==1.9.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.14.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120
extend-exclude = ["archive", "analysis_tmp.py"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = [
  # Used intentionally for OHLCV tuples (low = `l`) and other market-data idioms.
  "E741",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = [
  "-ra",
  "--cov=quant_trader_v5.sqlite_logger",
  "--cov=heartbeat",
  "--cov-report=term-missing",
  "--cov-fail-under=100",
]
